Add cmd for read bDone (Main.M<id>.bDone?).

+++ ecmcApp/src/motion/ecmcMotion.h
@@ -583,8 +583,8 @@ int getAxisTargetVel(int     axisIndex,
  *
  *  \todo  "TwinCAT syntax. Needs to be changed.\n
  */
-int getAxisDone(int     axisIndex,
-                double *value);
+int getAxisDone(int axisIndex,
+                int *value);
 
 /** \breif Get gear ration setting.\n
  *
